refactor: . Move src/*Renderer* to src/Renderer/
