+++ tests/TaskLineRenderer.test.ts
@@ -11,8 +11,8 @@ import { QueryLayoutOptions } from '../src/Layout/QueryLayoutOptions';
 import type { Task } from '../src/Task/Task';
 import { TaskRegularExpressions } from '../src/Task/Task';
 import { type TaskLayoutComponent, taskLayoutComponents } from '../src/Layout/TaskLayout';
-import type { TextRenderer } from '../src/TaskLineRenderer';
-import { TaskLineRenderer } from '../src/TaskLineRenderer';
+import type { TextRenderer } from '../src/Renderer/TaskLineRenderer';
+import { TaskLineRenderer } from '../src/Renderer/TaskLineRenderer';
 import { fromLine } from './TestingTools/TestHelpers';
 import { verifyWithFileExtension } from './TestingTools/ApprovalTestHelpers';
 import { TaskBuilder } from './TestingTools/TaskBuilder';
